modive/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.37.38
botocore==1.37.38
Django==4.2.21
django-cors-headers==4.4.0
djangorestframework==3.15.2
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
s3transfer==0.11.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==1.26.20
